requirements.txt
langchain
ipykernel
python-dotenv
langchain_community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
chromadb
sentence_transformers
langchain_huggingface
langchain_core
faiss-cpu
langchain_chroma 
langchain-ollama
langchain_groq
langchain_openai
openai
fastapi
uvicorn
langserve[all]
sse_starlette
pyserial
opencv-python
ultralytics
pydobotplus
